WIP refactoring GC tests
